+++ kinto/views/buckets.py
@@ -1,4 +1,7 @@
 from cliquet import resource
+from cliquet.events import ResourceChanged, ACTIONS
+from pyramid.events import subscriber
+
 from kinto.views import NameGenerator
 
 
@@ -8,7 +11,7 @@ class BucketSchema(resource.ResourceSchema):
 
 
 @resource.register(name='bucket',
-                   collection_methods=('GET', 'POST'),
+                   collection_methods=('GET', 'POST', 'DELETE'),
                    collection_path='/buckets',
                    record_path='/buckets/{{id}}')
 class Bucket(resource.ProtectedResource):
@@ -23,12 +26,20 @@ class Bucket(resource.ProtectedResource):
         # Buckets are not isolated by user, unlike Cliquet resources.
         return ''
 
-    def delete(self):
-        result = super(Bucket, self).delete()
+
+@subscriber(ResourceChanged,
+            for_resources=('bucket',),
+            for_actions=(ACTIONS.DELETE,))
+def on_buckets_deleted(event):
+    """Some buckets were deleted, delete sub-resources.
+    """
+    storage = event.request.registry.storage
+
+    for change in event.impacted_records:
+        bucket = change['old']
+        parent_id = '/buckets/%s' % bucket['id']
 
         # Delete groups.
-        storage = self.model.storage
-        parent_id = '/buckets/%s' % self.record_id
         storage.delete_all(collection_id='group',
                            parent_id=parent_id,
                            with_deleted=False)
@@ -36,20 +47,18 @@ class Bucket(resource.ProtectedResource):
                               parent_id=parent_id)
 
         # Delete collections.
-        deleted = storage.delete_all(collection_id='collection',
-                                     parent_id=parent_id,
-                                     with_deleted=False)
+        deleted_collections = storage.delete_all(collection_id='collection',
+                                                 parent_id=parent_id,
+                                                 with_deleted=False)
         storage.purge_deleted(collection_id='collection',
                               parent_id=parent_id)
 
         # Delete records.
-        id_field = self.model.id_field
-        for collection in deleted:
-            parent_id = '/buckets/%s/collections/%s' % (self.record_id,
-                                                        collection[id_field])
+        for collection in deleted_collections:
+            parent_id = '/buckets/%s/collections/%s' % (bucket['id'],
+                                                        collection['id'])
             storage.delete_all(collection_id='record',
                                parent_id=parent_id,
                                with_deleted=False)
-            storage.purge_deleted(collection_id='record', parent_id=parent_id)
-
-        return result
+            storage.purge_deleted(collection_id='record',
+                                  parent_id=parent_id)