+++ kinto/views/collections.py
@@ -1,7 +1,9 @@
 import colander
 import jsonschema
 from cliquet import resource
+from cliquet.events import ResourceChanged, ACTIONS
 from jsonschema import exceptions as jsonschema_exceptions
+from pyramid.events import subscriber
 
 from kinto.views import NameGenerator
 
@@ -34,7 +36,7 @@ class CollectionSchema(resource.ResourceSchema):
 
 
 @resource.register(name='collection',
-                   collection_methods=('GET', 'POST'),
+                   collection_methods=('GET', 'POST', 'DELETE'),
                    collection_path='/buckets/{{bucket_id}}/collections',
                    record_path='/buckets/{{bucket_id}}/collections/{{id}}')
 class Collection(resource.ProtectedResource):
@@ -50,16 +52,21 @@ class Collection(resource.ProtectedResource):
         parent_id = '/buckets/%s' % bucket_id
         return parent_id
 
-    def delete(self):
-        result = super(Collection, self).delete()
 
-        # Delete records.
-        storage = self.model.storage
-        parent_id = '%s/collections/%s' % (self.model.parent_id,
-                                           self.record_id)
+@subscriber(ResourceChanged,
+            for_resources=('collection',),
+            for_actions=(ACTIONS.DELETE,))
+def on_collections_deleted(event):
+    """Some collections were deleted, delete records.
+    """
+    storage = event.request.registry.storage
+
+    for change in event.impacted_records:
+        collection = change['old']
+        parent_id = '/buckets/%s/collections/%s' % (event.payload['bucket_id'],
+                                                    collection['id'])
         storage.delete_all(collection_id='record',
                            parent_id=parent_id,
                            with_deleted=False)
-        storage.purge_deleted(collection_id='record', parent_id=parent_id)
-
-        return result
+        storage.purge_deleted(collection_id='record',
+                              parent_id=parent_id)